*Tablet visning får desktop-ish dropdown meny. *Diverse tweaks og fixes på styling.
