+++ src/index.less
@@ -33,7 +33,8 @@ body {
 }
 
 .media-md-tablet,
-.media-lg-desktop {
+.media-lg-desktop,
+.media-tablet-desktop {
     display: none !important;
 }
 
@@ -42,7 +43,8 @@ body {
         display: none !important;
     }
 
-    .media-md-tablet {
+    .media-md-tablet,
+    .media-tablet-desktop {
         @media screen and (-ms-high-contrast: active),
             (-ms-high-contrast: none) {
             display: inline !important;
@@ -62,15 +64,15 @@ body {
 }
 
 @media (min-width: @desktopLimitWidth) {
-    .media-mobil-tablet {
-        display: none !important;
-    }
-
     .media-lg-desktop {
         @media screen and (-ms-high-contrast: active),
             (-ms-high-contrast: none) {
             display: inline !important;
         }
         display: initial !important;
+    }
+
+    .media-md-tablet {
+        display: none !important;
     }
 }
