+++ src/komponenter/header/meny/Desktopmeny.less
@@ -9,14 +9,15 @@
     &__content {
         width: 100%;
         margin: 0 auto;
-        padding: 0 1rem;
+        padding: 0;
     }
 
     &__elementer {
         display: flex;
         align-items: center;
         flex-wrap: wrap;
-        justify-content: space-around;
+        justify-content: space-between;
+        padding: 0;
     }
 
     &__nav-brand {
@@ -60,13 +61,8 @@
     }
 
     @media (min-width: @desktopLimitWidth) {
-        &__elementer {
-            justify-content: space-between;
-        }
-
         &__content {
             width: @desktopLimitWidth;
-            padding: 0;
         }
 
         .minsidelenke-varselbjelle {
