+++ src/styling-variabler.less
@@ -14,3 +14,5 @@
 
 //set common max-heigh for buttons and form-elements, to easily align-items on row.
 @maxheight-menuitem: 40px;
+
+@padding-horizontal: 3rem;
